requirements.txt
numpy
pandas
scipy
scikit-learn
PyYAML
matplotlib
torch
yfinance
